Move conf.utils to _utils and hoist imports
Rather than making someone import from documenteer.conf.utils to get
utilities, let's use the documenteer.conf namespace be used for these
utilities, and then rename the utils module for _utils so it won't be
directly used.
