+++ docs/dev/api/documenteer.conf.rst
@@ -2,4 +2,4 @@
 documenteer.conf
 ################
 
-.. automodapi:: documenteer.conf.utils
+.. automodapi:: documenteer.conf
