+++ src/documenteer/conf/__init__.py
@@ -1,0 +1,6 @@
+from ._utils import get_asset_path, get_template_dir
+
+__all__ = [
+    "get_asset_path",
+    "get_template_dir",
+]
